move definitions recommended by review
